+++ semcore/dropdown-menu/__tests__/index.test.tsx
@@ -2,14 +2,19 @@ import React from 'react';
 import { snapshot } from '@semcore/testing-utils/snapshot';
 import Button from '@semcore/button';
 import { expect, test, describe, beforeEach, vi } from '@semcore/testing-utils/vitest';
-import { cleanup, render, fireEvent, act, userEvent } from '@semcore/testing-utils/testing-library';
+import {
+  cleanup,
+  render,
+  fireEvent,
+  act,
+  userEvent,
+  waitFor,
+} from '@semcore/testing-utils/testing-library';
 import { axe } from '@semcore/testing-utils/axe';
 
 import DropdownMenu from '../src';
-import { getFocusableIn } from '@semcore/utils/lib/focus-lock/getFocusableIn';
 import { Box } from '@semcore/flex-box';
 import { ButtonTrigger } from '@semcore/base-trigger';
-import DesktopIconM from '@semcore/icon/Desktop/m';
 
 describe('DropdownMenu', () => {
   beforeEach(cleanup);
@@ -138,6 +143,7 @@ describe('DropdownMenu', () => {
 
     await userEvent.keyboard('[Tab]');
     await userEvent.keyboard('[Enter]');
+    await new Promise((resolve) => setTimeout(resolve, 500));
     await userEvent.keyboard('[ArrowRight]');
     await userEvent.keyboard('[Enter]');
 
@@ -224,12 +230,16 @@ describe('DropdownMenu', () => {
 
     await userEvent.keyboard('[Tab]');
     await userEvent.keyboard('[Enter]'); // open
+    await new Promise((resolve) => setTimeout(resolve, 500));
     await userEvent.keyboard('[Escape]'); // close
     await userEvent.keyboard('[Enter]'); // open
+    await new Promise((resolve) => setTimeout(resolve, 500));
     await userEvent.keyboard('[Enter]'); // click on the first item and close // 1
     await userEvent.keyboard('[Enter]'); // open
+    await new Promise((resolve) => setTimeout(resolve, 500));
     await userEvent.keyboard('[Enter]'); // click on the first item and close // 2
     await userEvent.keyboard('[Enter]'); // open
+    await new Promise((resolve) => setTimeout(resolve, 500));
     await userEvent.keyboard('[Enter]'); // click on the first item and close // 3
 
     expect(spy).toHaveBeenCalledTimes(3);
@@ -255,10 +265,42 @@ describe('DropdownMenu', () => {
 
     await userEvent.keyboard('[Tab]');
     await userEvent.keyboard('[Enter]');
+    await new Promise((resolve) => setTimeout(resolve, 500));
     await userEvent.keyboard('[ArrowDown]');
     await userEvent.keyboard('[Enter]');
 
     expect(spy).toHaveBeenCalledOnce();
+  });
+
+  test.sequential('Should close by second click on trigger', async ({ expect }) => {
+    const Component = () => {
+      return (
+        <DropdownMenu>
+          <DropdownMenu.Trigger tag='button' data-testid='dd-button-trigger'>
+            Trigger
+          </DropdownMenu.Trigger>
+          <DropdownMenu.Menu data-testid='dd-menu'>
+            <DropdownMenu.Item>Item 1</DropdownMenu.Item>
+            <DropdownMenu.Item>Item 2</DropdownMenu.Item>
+            <DropdownMenu.Item>Item 3</DropdownMenu.Item>
+          </DropdownMenu.Menu>
+        </DropdownMenu>
+      );
+    };
+    const { getByTestId } = render(<Component />);
+
+    await userEvent.keyboard('[Tab]');
+    await userEvent.keyboard('[Enter]');
+
+    const element = getByTestId('dd-menu');
+
+    expect(element).toBeInTheDocument();
+
+    await userEvent.click(getByTestId('dd-button-trigger'));
+
+    await waitFor(() => {
+      expect(element).not.toBeInTheDocument();
+    });
   });
 
   describe.sequential('opens nested menu', () => {
@@ -285,6 +327,7 @@ describe('DropdownMenu', () => {
 
       await userEvent.keyboard('[Tab]');
       await userEvent.keyboard('[Enter]');
+      await new Promise((resolve) => setTimeout(resolve, 500));
       await userEvent.keyboard('[ArrowDown]');
       await userEvent.keyboard('[Enter]');
       expect(getByTestId('item-2-2')).toBeTruthy();
@@ -312,6 +355,7 @@ describe('DropdownMenu', () => {
 
       await userEvent.keyboard('[Tab]');
       await userEvent.keyboard('[Enter]');
+      await new Promise((resolve) => setTimeout(resolve, 500));
       await userEvent.keyboard('[ArrowDown]');
       await userEvent.keyboard('[ArrowRight]');
       expect(getByTestId('item-2-2')).toBeTruthy();
@@ -341,6 +385,7 @@ describe('DropdownMenu', () => {
 
     await userEvent.keyboard('[Tab]');
     await userEvent.keyboard('[Enter]');
+    await new Promise((resolve) => setTimeout(resolve, 500));
     expect(getByTestId('dd-menu-item-1')).toHaveFocus();
 
     await userEvent.keyboard('[ArrowDown]');