+++ stories/components/base-trigger/docs/BaseTrigger.docs.stories.tsx
@@ -1,0 +1,13 @@
+import type { Meta } from '@storybook/react';
+
+import LinkTriggerExample from './examples/linktrigger';
+
+const meta = {
+  title: 'Components/BaseTrigger/Documentation',
+};
+
+export const LinkTrigger = {
+  render: LinkTriggerExample,
+};
+
+export default meta;