+++ semcore/button/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [5.40.1] - 2024-12-05
+
+### Removed
+
+- Unnecessary `aria-disabled` on disabled button.
+
+## [5.40.0] - 2024-11-29
+
+### Changed
+
+- Version minor update due to children dependencies update (`@semcore/flex-box` [5.38.2 ~> 5.39.0], `@semcore/utils` [4.43.3 ~> 4.44.0]).
+
 ## [5.39.4] - 2024-11-22
 
 ### Changed